requirements.txt
Flask==3.1.3
Requests==2.32.5
gunicorn
gevent
fpdf2
staticmap
Pillow